Move assert_matches to planned stable path
