+++ compiler/rustc_data_structures/src/lib.rs
@@ -42,8 +42,11 @@
 // have to worry about it being moved to a different module in std during stabilization.
 // FIXME(#151359): Remove this when `feature(assert_matches)` is stable in stage0.
 // (This doesn't necessarily need to be fixed during the beta bump itself.)
+#[cfg(bootstrap)]
 pub use std::assert_matches::{assert_matches, debug_assert_matches};
 use std::fmt;
+#[cfg(not(bootstrap))]
+pub use std::{assert_matches, debug_assert_matches};
 
 pub use atomic_ref::AtomicRef;
 pub use ena::{snapshot_vec, undo_log, unify};
