+++ library/alloctests/tests/str.rs
@@ -1,6 +1,6 @@
 #![allow(invalid_from_utf8)]
 
-use std::assert_matches::assert_matches;
+use std::assert_matches;
 use std::borrow::Cow;
 use std::cmp::Ordering::{Equal, Greater, Less};
 use std::str::{from_utf8, from_utf8_unchecked};
