+++ library/alloctests/tests/string.rs
@@ -1,10 +1,9 @@
-use std::assert_matches::assert_matches;
 use std::borrow::Cow;
 use std::cell::Cell;
 use std::collections::TryReserveErrorKind::*;
 use std::ops::Bound::*;
 use std::ops::{Bound, RangeBounds};
-use std::{panic, str};
+use std::{assert_matches, panic, str};
 
 pub trait IntoCow<'a, B: ?Sized>
 where
