+++ tests/ui-fulldeps/rustc_public/check_abi.rs
@@ -25,7 +25,7 @@ use rustc_public::mir::mono::Instance;
 use rustc_public::target::MachineInfo;
 use rustc_public::ty::{AdtDef, RigidTy, Ty, TyKind};
 use rustc_public::{CrateDef, CrateItem, CrateItems, ItemKind};
-use std::assert_matches::assert_matches;
+use std::assert_matches;
 use std::collections::HashSet;
 use std::convert::TryFrom;
 use std::io::Write;
