+++ tests/ui-fulldeps/rustc_public/check_intrinsics.rs
@@ -24,7 +24,7 @@ use rustc_public::mir::mono::{Instance, InstanceKind};
 use rustc_public::mir::visit::{Location, MirVisitor};
 use rustc_public::mir::{LocalDecl, Terminator, TerminatorKind};
 use rustc_public::ty::{FnDef, GenericArgs, RigidTy, TyKind};
-use std::assert_matches::assert_matches;
+use std::assert_matches;
 use std::convert::TryFrom;
 use std::io::Write;
 use std::ops::ControlFlow;
