+++ tests/ui/macros/assert-matches-macro-msg.rs
@@ -6,7 +6,7 @@
 
 #![feature(assert_matches)]
 
-use std::assert_matches::assert_matches;
+use std::assert_matches;
 
 fn main() {
     assert_matches!(1 + 1, 3, "1 + 1 definitely should be 3");
